requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.3.0
mysqlclient==2.2.7
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
